decoder/requirements.txt
git+https://github.com/lbryio/lbryschema.git#egg=lbryschema
python-bitcoinrpc==0.1
flask